CUBA 7.0 migration
