+++ build.gradle
@@ -1,6 +1,6 @@
 
 buildscript {
-    ext.cubaVersion = '6.10.2'
+    ext.cubaVersion = '7.0.0.BETA1'
     repositories {
 
         mavenLocal()
@@ -46,7 +46,6 @@ def webModule = project(":${modulePrefix}-web")
 def servletApi = 'javax.servlet:javax.servlet-api:3.1.0'
 
 
-apply(plugin: 'idea')
 apply(plugin: 'cuba')
 
 
@@ -55,7 +54,7 @@ apply(plugin: 'cuba')
 cuba {
     artifact {
         group = 'de.diedavids.cuba.runtimediagnose'
-        version = "1.2.0"
+        version = "1.3.0"
         isSnapshot = false
     }
     tomcat {
@@ -150,13 +149,11 @@ test.finalizedBy(project.tasks.coberturaCheck)
 configure([globalModule, coreModule, guiModule, webModule]) {
     apply(plugin: 'java')
     apply(plugin: 'maven')
-    apply(plugin: 'idea')
     apply(plugin: 'cuba')
 
     dependencies {
         testCompile('junit:junit:4.12')
-
-
+        compile('org.codehaus.groovy:groovy-all:2.5.4')
         testCompile('org.spockframework:spock-core:1.0-groovy-2.4')
         testCompile('org.springframework:spring-test:4.3.1.RELEASE')
         testRuntime "cglib:cglib-nodep:3.2.4"
@@ -238,7 +235,7 @@ configure(coreModule) {
 
     dependencies {
         compile(globalModule)
-        provided(servletApi)
+        compileOnly(servletApi)
         jdbc(hsql)
         testRuntime(hsql)
 
@@ -298,7 +295,7 @@ configure(webModule) {
     }
 
     dependencies {
-        provided(servletApi)
+        compileOnly(servletApi)
         compile(guiModule)
 
     }
@@ -362,9 +359,6 @@ task restart(dependsOn: ['stop', ":${modulePrefix}-core:deploy", ":${modulePrefi
     }
 }
 
-task wrapper(type: Wrapper) {
-    gradleVersion = '4.3.1'
-}
 
 apply from: 'extra.gradle'
 
@@ -372,3 +366,5 @@ apply from: 'extra.gradle'
 clean {
     delete = ['build/libs', 'build/tmp']
 }
+
+task buildUberJar(type: CubaUberJarBuilding) {}
