+++ modules/core/db/init/hsql/20.create-db.sql
@@ -1,4 +1,4 @@
 -- begin DDCRD_DIAGNOSE_EXECUTION_LOG
-alter table DDCRD_DIAGNOSE_EXECUTION_LOG add constraint FK_DDCRD_DIAGNOSE_EXECUTION_LOG_EXECUTION_RESULT_FILE foreign key (EXECUTION_RESULT_FILE_ID) references SYS_FILE(ID)^
-create index IDX_DDCRD_DIAGNOSE_EXECUTION_LOG_EXECUTION_RESULT_FILE on DDCRD_DIAGNOSE_EXECUTION_LOG (EXECUTION_RESULT_FILE_ID)^
+alter table DDCRD_DIAGNOSE_EXECUTION_LOG add constraint FK_DDCRD_DIAGNOSE_EXECUTION_LOG_ON_EXECUTION_RESULT_FILE foreign key (EXECUTION_RESULT_FILE_ID) references SYS_FILE(ID)^
+create index IDX_DDCRD_DIAGNOSE_EXECUTION_LOG_ON_EXECUTION_RESULT_FILE on DDCRD_DIAGNOSE_EXECUTION_LOG (EXECUTION_RESULT_FILE_ID)^
 -- end DDCRD_DIAGNOSE_EXECUTION_LOG
