+++ modules/core/db/update/hsql/18/181119-2-createDiagnoseExecutionLog.sql
@@ -1,0 +1,2 @@
+alter table DDCRD_DIAGNOSE_EXECUTION_LOG add constraint FK_DDCRD_DIAGNOSE_EXECUTION_LOG_ON_EXECUTION_RESULT_FILE foreign key (EXECUTION_RESULT_FILE_ID) references SYS_FILE(ID);
+create index IDX_DDCRD_DIAGNOSE_EXECUTION_LOG_ON_EXECUTION_RESULT_FILE on DDCRD_DIAGNOSE_EXECUTION_LOG (EXECUTION_RESULT_FILE_ID);
