+++ modules/web/src/de/diedavids/cuba/runtimediagnose/web-app.properties
@@ -39,7 +39,7 @@ cuba.webContextName = runtime-diagnose
 # Web session inactivity timeout in seconds
 cuba.httpSessionExpirationTimeoutSec = 1800
 
-cuba.web.theme = halo
+cuba.web.theme = hover
 cuba.web.foldersPaneEnabled=false
 cuba.availableLocales = English|en
 cuba.localeSelectVisible = false
