+++ modules/web/src/de/diedavids/cuba/runtimediagnose/web/screens/jpql/JpqlConsole.groovy
@@ -1,60 +1,52 @@
 package de.diedavids.cuba.runtimediagnose.web.screens.jpql
 
+import com.haulmont.bali.util.ParamsMap
+import com.haulmont.cuba.core.global.RemoteException
+import com.haulmont.cuba.gui.UiComponents
+import com.haulmont.cuba.gui.components.Button
+import com.haulmont.cuba.gui.components.SourceCodeEditor
+import com.haulmont.cuba.gui.components.actions.BaseAction
+import de.diedavids.cuba.runtimediagnose.db.DbDiagnoseService
+import de.diedavids.cuba.runtimediagnose.diagnose.DiagnoseType
+import de.diedavids.cuba.runtimediagnose.web.screens.console.ConsoleWindow
+
+import javax.inject.Inject
 
 import static com.haulmont.cuba.gui.WindowManager.OpenType.DIALOG
 import static com.haulmont.cuba.gui.components.Frame.NotificationType.ERROR
 import static com.haulmont.cuba.gui.components.Frame.NotificationType.WARNING
 
-import com.haulmont.bali.util.ParamsMap
-import com.haulmont.cuba.core.global.RemoteException
-import com.haulmont.cuba.gui.components.AbstractWindow
-import com.haulmont.cuba.gui.components.Button
-import com.haulmont.cuba.gui.components.SourceCodeEditor
-import com.haulmont.cuba.gui.components.actions.BaseAction
-import com.haulmont.cuba.gui.xml.layout.ComponentsFactory
-import de.diedavids.cuba.runtimediagnose.db.DbDiagnoseService
-import de.diedavids.cuba.runtimediagnose.diagnose.DiagnoseType
-import de.diedavids.cuba.runtimediagnose.web.screens.console.ConsoleFrame
-
-import javax.inject.Inject
-
-
-class JpqlConsole extends AbstractWindow {
+class JpqlConsole extends ConsoleWindow {
 
     @Inject
     DbDiagnoseService dbDiagnoseService
 
     @Inject
-    ConsoleFrame consoleFrame
-
-    @Inject
-    ComponentsFactory componentsFactory
+    UiComponents uiComponents
 
     @Override
     void init(Map<String, Object> params) {
         super.init(params)
-
-        consoleFrame = openFrame(
-                consoleFrame,
-                'console-frame',
-                ParamsMap.of('diagnoseType', DiagnoseType.JPQL)
-        )
-
         addSqlQueryButton()
     }
 
+    @Override
+    DiagnoseType getDiagnoseType() {
+        return DiagnoseType.JPQL
+    }
+
     void addSqlQueryButton() {
-        Button showSqlBtn = componentsFactory.createComponent(Button)
+        Button showSqlBtn = uiComponents.create(Button)
         showSqlBtn.with {
             caption = getMessage('generateSql')
             setAction(new BaseAction('showSqlDialog').withHandler { e -> showSqlDialog() })
         }
 
-        consoleFrame.getComponent('consoleInstrumentalPanel').add(showSqlBtn)
+        getComponent('consoleInstrumentalPanel').add(showSqlBtn)
     }
 
     void showSqlDialog() {
-        SourceCodeEditor codeEditor = consoleFrame.getComponent('console') as SourceCodeEditor
+        SourceCodeEditor codeEditor = getComponent('console') as SourceCodeEditor
 
         if (codeEditorIsEmpty(codeEditor)) {
             showNotification(getMessage('noScriptDefined'), WARNING)
