+++ modules/web/src/de/diedavids/cuba/runtimediagnose/web/screens/sql/sql-console.xml
@@ -2,10 +2,7 @@
 <window xmlns="http://schemas.haulmont.com/cuba/window.xsd"
         caption="msg://caption"
         class="de.diedavids.cuba.runtimediagnose.web.screens.sql.SqlConsole"
-        messagesPack="de.diedavids.cuba.runtimediagnose.web.screens.sql">
-    <layout height="100%"
-            width="100%"
-            expand="consoleFrame">
-        <frame id="consoleFrame" screen="console-frame"/>
+        extends="/de/diedavids/cuba/runtimediagnose/web/screens/console/console-frame.xml">
+    <layout>
     </layout>
 </window>
