+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'cuba-component-runtime-diagnose'
-include(':runtime-diagnose-global', ':runtime-diagnose-core', ':runtime-diagnose-gui', ':runtime-diagnose-web')
-project(':runtime-diagnose-global').projectDir = new File(settingsDir, 'modules/global')
-project(':runtime-diagnose-core').projectDir = new File(settingsDir, 'modules/core')
-project(':runtime-diagnose-gui').projectDir = new File(settingsDir, 'modules/gui')
-project(':runtime-diagnose-web').projectDir = new File(settingsDir, 'modules/web')
+def modulePrefix = 'runtime-diagnose'
+include(":${modulePrefix}-global", ":${modulePrefix}-core", ":${modulePrefix}-gui", ":${modulePrefix}-web")
+project(":${modulePrefix}-global").projectDir = new File(settingsDir, "modules/global")
+project(":${modulePrefix}-core").projectDir = new File(settingsDir, "modules/core")
+project(":${modulePrefix}-gui").projectDir = new File(settingsDir, "modules/gui")
+project(":${modulePrefix}-web").projectDir = new File(settingsDir, "modules/web")
